requirements.txt
# Streamlit UI
streamlit>=1.30.0

# ML / YOLO
torch>=2.0.0
ultralytics>=8.2.0

# Data Handling
numpy>=1.26.0
pandas>=2.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.1

# Visualization
matplotlib>=3.8.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0

# Machine Learning
scikit-learn>=1.3.0streamlit-webrtc
PyYAML
